requirements.txt
streamlit
rasterio
numpy
Pillow
matplotlib
streamlit-image-select
folium
streamlit-folium
